+++ swarm/src/behaviour.rs
@@ -117,41 +117,6 @@ pub(crate) type THandlerOutEvent<THandler> =
 ///   }
 /// }
 /// ```
-///
-/// Struct members that don't implement [`NetworkBehaviour`] must be annotated with
-/// `#[behaviour(ignore)]`.
-///
-/// ``` rust
-/// # use libp2p::identify::{Identify, IdentifyEvent};
-/// # use libp2p::ping::{Ping, PingEvent};
-/// # use libp2p::NetworkBehaviour;
-/// #[derive(NetworkBehaviour)]
-/// #[behaviour(out_event = "Event")]
-/// struct MyBehaviour {
-///   identify: Identify,
-///   ping: Ping,
-///
-///   #[behaviour(ignore)]
-///   some_string: String,
-/// }
-/// #
-/// # enum Event {
-/// #   Identify(IdentifyEvent),
-/// #   Ping(PingEvent),
-/// # }
-/// #
-/// # impl From<IdentifyEvent> for Event {
-/// #   fn from(event: IdentifyEvent) -> Self {
-/// #     Self::Identify(event)
-/// #   }
-/// # }
-/// #
-/// # impl From<PingEvent> for Event {
-/// #   fn from(event: PingEvent) -> Self {
-/// #     Self::Ping(event)
-/// #   }
-/// # }
-/// ```
 pub trait NetworkBehaviour: 'static {
     /// Handler for all the protocols the network behaviour supports.
     type ConnectionHandler: IntoConnectionHandler;